Move CAVP/SHAVS tests into their own crate.
Stop including these large test input files in the published *ring* crate,
as the *ring* crate is too large.
